add-logic-for-patch

+++ src/app.ts
@@ -1,13 +1,13 @@
 import express, { Request, Response } from "express";
 import path from "path";
 import config from "./config";
-const { products, carts } = config;
-import { readdir, readFile, writeFile, appendFile } from "fs/promises";
+import { readdir, readFile, writeFile, unlink } from "fs/promises";
 import type { Cart, Product } from "./__tests__/e2e/e2e-types";
 import { v4 as uuidv4 } from "uuid";
 import { validate } from "./cart/api/validation";
-import { constants } from "buffer";
+import { existsSync } from "fs";
 
+const { products, carts } = config;
 const app = express();
 
 // Some nice middleware :)
@@ -70,39 +70,64 @@ app.post("/api/carts/", (req: Request, res: Response) => {
 });
 
 app.patch("/api/carts/:id", async (req: Request, res: Response) => {
-
-  const product = validate({
-    ...req.body,
-    quantity: Number(req.body.quantity),
-    price: Number(req.body.price)
-  });
-
-  if (!product.success) {
-    res.status(400).json({ error: "Invalid request" });
+  if (!req.body) {
+    res.status(400).json({ error: "Bad Request" });
   } else {
-    const cartPath = path.join(carts.db, req.params.id);
+    const product = validate({
+      ...req.body,
+      quantity: Number(req.body.quantity),
+      price: Number(req.body.price),
+    });
 
-    const products = JSON.parse(await readFile(cartPath, "utf-8")) as Product[];
-    products.push(product.data);
-    await writeFile(cartPath, JSON.stringify(products), "utf8");
+    if (!product.success) {
+      res.status(400).json({ error: "Invalid request" });
+    } else {
+      const cartPath = path.join(carts.db, req.params.id);
+      if (!existsSync(cartPath)) {
+        res.status(400).json("Bad Request");
+      } else {
+        const products = JSON.parse(
+          await readFile(cartPath, "utf-8"),
+        ) as Product[];
+        products.push(product.data);
+        await writeFile(cartPath, JSON.stringify(products), "utf8");
 
-    res.status(204).json("Added product to cart: " + product.data.id);
+        res.status(204).json("Added product to cart: " + product.data.id);
+      }
+    }
   }
 });
 
 app.get("/api/carts/:id", async (req: Request, res: Response) => {
   try {
     const filePath = path.join(carts.db, req.params.id);
-    const products = JSON.parse(await readFile(filePath, "utf-8")) as Product[];
+    if (!existsSync(filePath)) {
+      res.status(404).json("This cart doesn't exist");
+    } else {
+      const products = JSON.parse(
+        await readFile(filePath, "utf-8"),
+      ) as Product[];
 
-    const cart: Cart = {
-      id: req.params.id,
-      products,
-    };
+      const cart: Cart = {
+        id: req.params.id,
+        products,
+      };
 
-    res.json(cart);
+      res.json(cart);
+    }
   } catch (err) {
     console.error(err);
+  }
+});
+
+app.delete("/api/carts/:id", async (req: Request, res: Response) => {
+  const cartPath = path.join(carts.db, req.params.id);
+
+  if (existsSync(cartPath)) {
+    await unlink(cartPath);
+    res.status(204).json();
+  } else {
+    res.status(204).json("Cart doesn't exists");
   }
 });
